Move code only used for testing/benchmarking into tests
